+++ shared/actions/kbfs.desktop.js
@@ -1,19 +1,19 @@
 // @flow
-import * as ConfigGen from '../config-gen'
-import * as KBFSGen from '../kbfs-gen'
-import * as Constants from '../../constants/config'
-import * as Saga from '../../util/saga'
-import * as RPCTypes from '../../constants/types/flow-types'
+import * as ConfigGen from './config-gen'
+import * as KBFSGen from './kbfs-gen'
+import * as Constants from '../constants/config'
+import * as Saga from '../util/saga'
+import * as RPCTypes from '../constants/types/flow-types'
 import electron, {shell} from 'electron'
 import fs from 'fs'
 import path from 'path'
-import {ExitCodeFuseKextPermissionError} from '../../constants/favorite'
+import {ExitCodeFuseKextPermissionError} from '../constants/favorite'
 import {delay} from 'redux-saga'
 import {execFile} from 'child_process'
-import {folderTab} from '../../constants/tabs'
-import {isLinux, isWindows} from '../../constants/platform'
-import {navigateTo, switchTo} from '../route-tree'
-import type {TypedState} from '../../constants/reducer'
+import {folderTab} from '../constants/tabs'
+import {isLinux, isWindows} from '../constants/platform'
+import {navigateTo, switchTo} from './route-tree'
+import type {TypedState} from '../constants/reducer'
 
 // pathToURL takes path and converts to (file://) url.
 // See https://github.com/sindresorhus/file-url
