--- shared/actions/kbfs/index.js
@@ -1,3 +1,0 @@
-// @flow
-import saga from './index.platform'
-export default saga
